fix mocking

+++ packages/trpc/server/routers/viewer/teams/managedEvents/managedEventManualReassign.schema.ts
@@ -1,8 +1,8 @@
 import { z } from "zod";
 
 export const ZManagedEventManualReassignInputSchema = z.object({
-  bookingId: z.number(),
-  teamMemberId: z.number(),
+  bookingId: z.number().int(),
+  teamMemberId: z.number().int(),
   reassignReason: z.string().optional(),
 });
